requirements.txt
Django>=4.1.7,<4.2
djangorestframework>=3.14.0,<3.15
django-cors-headers>=3.14.0,<3.15